WT-10 data is now received through the redux process
