+++ source/js/api/index.js
@@ -1,16 +1,25 @@
 import 'es6-promise';
 
+const moment = require('moment');
+const _ = require('lodash');
+
 function testAsync() {
   return new Promise(resolve => {
     setTimeout(() => {
-      const date = new Date();
-      let seconds = date.getSeconds();
-      let minutes = date.getMinutes();
+      const date = moment();
+      const currentDate = _.cloneDeep(date);
 
-      seconds = seconds < 10 ? `0${ seconds }` : seconds;
-      minutes = minutes < 10 ? `0${ minutes }` : minutes;
-
-      resolve(`Current time: ${ date.getHours() }:${ minutes }:${ seconds }`);
+      resolve({
+        xValues: [
+          date.subtract(5, 'days').format('MM/DD/YYYY'),
+          date.add(1, 'days').format('MM/DD/YYYY'),
+          date.add(1, 'days').format('MM/DD/YYYY'),
+          date.add(1, 'days').format('MM/DD/YYYY'),
+          date.add(1, 'days').format('MM/DD/YYYY'),
+          currentDate.format('MM/DD/YYYY'),
+        ],
+        yValues: [183, 185, 187, 184, 184, 187],
+      });
     }, (Math.random() * 1000) + 1000); // 1-2 seconds delay
   });
 }
